- refactor

+++ lib/features/user_role/admin/admin_home/atomic_ui/atoms/admin_notification_item.dart
@@ -1,5 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
+import 'package:employee_mangement/core/common/app_container_decoration.dart';
 import 'package:employee_mangement/core/utils/assets_manager.dart';
+import 'package:employee_mangement/core/widgets/build_alart_message.dart';
 import 'package:employee_mangement/core/widgets/custom_app_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -7,6 +9,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/svg.dart';
 
 import '../../../../../../core/helpers/app_spaces.dart';
+import '../../../../../../core/helpers/extention.dart';
 import '../../../../../../core/styles/colors.dart';
 import '../../../../../../core/styles/styles.dart';
 import '../../controllers/mange_employee_cubit/employee_cubit.dart';
@@ -27,71 +30,56 @@ class AdminNotificationItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Card(
-      color: Colors.white,
+    return Container(
       margin: const EdgeInsets.symmetric(vertical: 5.0, horizontal: 16.0),
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12.0),
-      ),
-      elevation: 3,
+      decoration: AppConatinerDecoration(),
       child: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text('Add New User Request'.tr(context: context),
-                    style: AppStylesManger.font14RegularBlack),
-                Text(
-                  "${"Name".tr(context: context)} : $name ",
-                  style: const TextStyle(
-                    fontSize: 12.0,
-                    fontWeight: FontWeight.bold,
-                    color: Colors.black87,
-                  ),
-                ),
-                const SizedBox(height: 4.0),
-                SizedBox(
-                  width: MediaQuery.of(context).size.width * 0.7,
-                  child: Text(
-                    "${"Mobile ".tr(context: context)} : $mobileId",
-                    style: TextStyle(
+            IntrinsicWidth(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text('Add New User Request'.tr(context: context),
+                      style: AppStylesManger.font14RegularBlack),
+                  Text(
+                    "${"Name".tr(context: context)} : $name ",
+                    style: const TextStyle(
                       fontSize: 12.0,
-                      color: Colors.grey[700],
                       fontWeight: FontWeight.bold,
-                      overflow: TextOverflow.ellipsis,
+                      color: Colors.black87,
                     ),
-                    maxLines: 1,
                   ),
-                ),
-                verticalSpace(10),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    CustomAppButton(
-                      height: 35.h,
-                      width: 100.w,
-                      textButton: 'Add'.tr(context: context),
-                      buttonColor: ColorsManger.primaryColor,
-                      onPressed: onTap,
-                    ),
-                    horizontalSpace(60),
-                    Container(
-                      padding: const EdgeInsets.all(8),
-                      decoration: BoxDecoration(
-                        gradient: LinearGradient(
-                            begin: Alignment.centerLeft,
-                            end: Alignment.centerRight,
-                            colors: [
-                              ColorsManger.primaryColor,
-                              const Color.fromARGB(255, 150, 144, 144),
-                            ]),
-                        borderRadius: BorderRadius.circular(14),
+                  horizontalSpace(8),
+                  SizedBox(
+                    width: MediaQuery.of(context).size.width * 0.7,
+                    child: Text(
+                      "${"Mobile ".tr(context: context)} : $mobileId",
+                      style: TextStyle(
+                        fontSize: 12.0,
+                        color: Colors.grey[700],
+                        fontWeight: FontWeight.bold,
+                        overflow: TextOverflow.ellipsis,
                       ),
-                      child: Row(
+                      maxLines: 1,
+                    ),
+                  ),
+                  verticalSpace(10),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      CustomAppButton(
+                        height: 35.h,
+                        width: 100.w,
+                        textButton: 'Add'.tr(context: context),
+                        buttonColor: ColorsManger.primaryColor,
+                        onPressed: onTap,
+                      ),
+                      Spacer(),
+                      Row(
                         spacing: 6.w,
                         children: [
                           Text(
@@ -99,44 +87,47 @@ class AdminNotificationItem extends StatelessWidget {
                               tr("dd MMM", context: context),
                               context.locale.toString(),
                             ).format(date),
-                            style: AppStylesManger.font16blackMedium.copyWith(
-                              color: ColorsManger.scaffoldBackgroundColor,
+                            style: AppStylesManger.font12RegularBlack.copyWith(
+                              color: Colors.black,
                             ),
                           ),
                           SvgPicture.asset(
                             Assets.assetsImagesCalenderAttendance,
+                            height: 12.h,
+                            width: 12.w,
                             colorFilter: ColorFilter.mode(
-                              ColorsManger.scaffoldBackgroundColor,
+                              Colors.black,
                               BlendMode.srcIn,
                             ),
                           ),
                         ],
                       ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-            Expanded(
-              child: Row(
-                spacing: 16,
-                children: [
-                  Flexible(
-                    child: Icon(
-                      Icons.notifications_active,
-                      color: ColorsManger.primaryColor,
-                    ),
+                    ],
                   ),
-                  Flexible(
-                    child: IconButton(
-                      onPressed: () {
+                ],
+              ),
+            ),
+            IntrinsicWidth(
+              child: Row(
+                children: [
+                  IconButton(
+                    onPressed: () {
+                      buildDeleteAlertDialog(
+                          message:
+                              "Are you sure you want to delete this request?"
+                                  .tr(context: context),
+                          context,
+                          title: "Delete Request".tr(context: context),
+                          onYes: () {
                         context
                             .read<EmployeeCubit>()
-                            .deleteAddAccountRequest(id: id);
-                      },
-                      icon:
-                          Icon(Icons.delete, color: ColorsManger.primaryColor),
-                    ),
+                            .deleteAddAccountRequest(id: id)
+                            .then((value) {
+                          context.pop();
+                        });
+                      });
+                    },
+                    icon: Icon(Icons.delete, color: ColorsManger.primaryColor),
                   ),
                 ],
               ),
